feat : rename common to shared
