+++ src/main/java/com/example/application/mfschemes/entities/MFSchemeNav.java
@@ -1,6 +1,6 @@
 package com.example.application.mfschemes.entities;
 
-import com.example.application.common.Auditable;
+import com.example.application.shared.Auditable;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
