+++ src/main/java/com/example/application/portfolio/entities/UserFolioDetails.java
@@ -1,7 +1,18 @@
 package com.example.application.portfolio.entities;
 
-import com.example.application.common.Auditable;
-import jakarta.persistence.*;
+import com.example.application.shared.Auditable;
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Index;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.SequenceGenerator;
+import jakarta.persistence.Table;
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
