+++ src/main/java/com/example/application/shared/config/AuditConfiguration.java
@@ -1,4 +1,4 @@
-package com.example.application.config;
+package com.example.application.shared.config;
 
 import java.util.Optional;
 import org.springframework.context.annotation.Bean;
